XML Sorting 1/2
We want to get rid of the technical debt of unsorted strings in the XML, as described in https://github.com/TheStarport/freelancer-bmod/issues/92.

I am writing tools to automate this process as much as possible. The updated XML will be part of an upcoming horror merge.

Also, added strings for Liberty pilots. TGA files are almost done and have received a reskin and have been created from reusable parts following the xcf template as much as possible. Feedback welcome.